paasio: move interfaces into a separate file
