+++ paasio/example.go
@@ -6,25 +6,6 @@ import (
 	"sync"
 )
 
-type ReadCounter interface {
-	io.Reader
-	// ReadCount returns the total number of bytes successfully read along
-	// with the total number of calls to Read().
-	ReadCount() (n int64, nops int)
-}
-
-type WriteCounter interface {
-	io.Writer
-	// WriteCount returns the total number of bytes successfully written along
-	// with the total number of calls to Write().
-	WriteCount() (n int64, nops int)
-}
-
-type ReadWriteCounter interface {
-	ReadCounter
-	WriteCounter
-}
-
 // NewWriteCounter returns an implementation of WriteCounter.  Calls to
 // w.Write() are not guaranteed to be synchronized.
 func NewWriteCounter(w io.Writer) (WriteCounter, error) {
